+++ src/index.ts
@@ -1,1 +1,2 @@
+export { LayoutProvider } from './context'
 export * from './components'