+++ src/interfaces/BaseProps.ts
@@ -1,0 +1,6 @@
+import type { Padding } from '../types/Padding'
+
+export interface BaseProps {
+  readonly spaces?: number
+  readonly padding?: Padding
+}