+++ CHANGELOG.md
@@ -1,1 +1,8 @@
 # native-layout-stack Changelog
+
+## 1.0.0 (Soon)
+
+### Features
+
+- added `Stack`, `Spacer` components
+- added `LayoutProvider`